+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Google provider updated to use the [Google Gen AI SDK](https://googleapis.github.io/python-genai/), which is now the recommended API for Gemini 2.0 models.
 - Task display: Print task progress every 5 seconds for 'plain' display mode.
+- Task display: Handle click on running samples tab when there is no transcript.
 
 ## v0.3.68 (19 February 2025)